Move Price to dumbs
