+++ src/cow-react/swap/dumb/TradeRates/index.tsx
@@ -4,7 +4,7 @@ import { InfoIcon } from 'components/InfoIcon'
 import { SUBSIDY_INFO_MESSAGE } from 'components/CowSubsidyModal/constants'
 import { useOpenModal } from 'state/application/hooks'
 import { ApplicationModal } from 'state/application/reducer'
-import { Price } from 'cow-react/swap/smart/Price'
+import { Price } from 'cow-react/swap/dumb/Price'
 import TradeGp from 'state/swap/TradeGp'
 import { INITIAL_ALLOWED_SLIPPAGE_PERCENT } from 'constants/index'
 import { RowSlippage } from 'components/swap/TradeSummary/RowSlippage'
