+++ src/custom/pages/Swap/SwapMod.tsx
@@ -49,7 +49,7 @@ import { useAtomValue } from 'jotai/utils'
 import { swapConfirmAtom } from 'cow-react/swap/state/swapConfirmAtom'
 import { SwapButton, SwapButtonContext } from 'cow-react/swap/dumb/SwapButton/SwapButton'
 import { RemoveRecipient } from 'cow-react/swap/smart/RemoveRecipient'
-import { Price } from 'cow-react/swap/smart/Price'
+import { Price } from 'cow-react/swap/dumb/Price'
 import { TradeBasicDetails } from 'cow-react/swap/smart/TradeBasicDetails'
 import { BottomGrouping } from 'cow-react/swap/dumb/styled'
 import { ArrowWrapperLoader } from 'components/ArrowWrapperLoader'
